refactor broker tests to reuse functions
